Configuration update: allow leaving stuff out
With the new extension addition, this would crash the plugin when users
update, since the plugin expects each config option to be present.

This is not the case anymore now, some config options have default
values which are used when they're missing from the config file.
e.g.: the list of enabled plugins is an empty table when not present in
the config.

Other options which are not marked as required can be left out
completely.